FORGOT TO COMMIT AGAIN
Lots has happened... there is a pretty big start on methods, calls and returns but they dont work correctly yet.
just fixed a bug with statements and that all works fine again.
scoped values are a bit broken but are currently being worked on.
i need to remember to commit :/